Rearranged folders, separated getData from view into separate class, removed redundant and unused code, renamed variables
