+++ Recipess-app/Views/ContentView/ViewComponents/SignInView.swift
@@ -1,41 +1,18 @@
 //
-//  ContentView.swift
+//  SignInView.swift
 //  Recipess-app
 //
-//  Created by A J on 2022-02-08.
+//  Created by Henrik Sjögren on 2022-06-07.
 //
 
 import SwiftUI
 
-struct ContentView: View {
-    @StateObject var loginModel = LoginViewModel()
-    @State var isLoading = false
-    
-    var body: some View {
-            if loginModel.isLoggedIn {
-                
-                UserSearchView()
-            } else {
-                SignIn(loginModel: loginModel)
-            }
-        }
-    
-    func startFakeNetworkCall() {
-        isLoading = true
-        DispatchQueue.main.asyncAfter(deadline: .now() + 1.5) {
-            isLoading = false
-        }
-    }
-    
-}
-
-struct SignIn: View{
-    
+struct SignInView: View {
     @State private var email = ""
     @State private var password = ""
     @ObservedObject var loginModel : LoginViewModel
     
-    var body: some View{
+    var body: some View {
         NavigationView{
             VStack(alignment: .leading, spacing: 20, content: {
                 Image("foodlogo")
@@ -43,7 +20,6 @@ struct SignIn: View{
                     .aspectRatio(contentMode: .fit)
                 
                 VStack{
-                    
                     HStack{
                         Image(systemName: "envelope")
                             .font(.title2)
@@ -74,11 +50,9 @@ struct SignIn: View{
                     .cornerRadius(30)
                     .padding(.horizontal, 20)
                     
-                    
                     VStack{
                         Button {
                             loginModel.signIn(email: email, password: password)
-                            
                         } label: {
                             Text("Login")
                                 .fontWeight(.bold)
@@ -90,21 +64,8 @@ struct SignIn: View{
                                 .shadow(radius: 10)
                         }
                         .padding()
+                        .padding(.bottom,50)
                         
-                        /*
-                        NavigationLink(destination: GuestSearchView(), label: {
-                            Text("Guest")
-                                .fontWeight(.bold)
-                                .foregroundColor(.white)
-                                .padding()
-                                .padding(.horizontal, 25)
-                                .background(Color("ColorRed"))
-                                .cornerRadius(30)
-                                .shadow(radius: 10)
-                            
-                        })
-                        */
-                            .padding(.bottom,50)
                         HStack{
                             NavigationLink(destination: RegisterUserView(), label:{
                                 Text("Register")
@@ -117,7 +78,6 @@ struct SignIn: View{
                 }
             })
                 .navigationTitle("Sign in")
-            // .navigationViewStyle(StackNavigationViewStyle())
         }
     }
     private func errorLogin() -> Alert {
@@ -125,22 +85,10 @@ struct SignIn: View{
     }
 }
 
-struct ContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        ContentView()
-    }
-}
 
-struct LoadingView: View {
-    var body: some View {
-        ZStack{
-            Color(.systemBackground)
-                .ignoresSafeArea()
-                .opacity(0.5)
-            
-            ProgressView()
-                .progressViewStyle(CircularProgressViewStyle(tint: .red))
-                .scaleEffect(3)
-        }
-    }
-}
+
+//struct SignInView_Previews: PreviewProvider {
+//    static var previews: some View {
+//        SignInView()
+//    }
+//}
